repair install path for appstream data
it occurs to me that the subdir simply got lost when changing the destination
from a hardcoded value to the variable, making the appdata file
end up in $prefix/share rather than $prefix/share/appdata. as far as I
can tell this was the only change gone wrong out of related changes to
the applications/15.04 branches.

CCMAIL: heirecka@exherbo.org