Deactivate pause button when no game is running
Before, the pause button was always activated even when no game was
running. This lead to undesired behavior, e.g. making it necessary to
press the pause button twice to pause the game.

With this fix, the pause button is deactivated on startup. The button
gets activated when a game is started and when the button is pressed, it
checks whether there is a running game.

BUG: 264088
REVIEW: 126200