Fix Coverty #76262: Check for NULL pointer
KBlocksAIPlayer::planning() did not check if pointer mpPlanner is NULL.
This check has been added and if mpPlanner is NULL a warning is issued
and the function left.

I have tested this case by commenting out the creation of the Planner.
As expected, the AI issues warnings that no planner has been set and
behaves extremely stupid (all Blocks just drop without moving or
rotating). No crash is encountered.