Fix Coverity #76274: Uninitialized members
In KBlocksGameLogic constructor mInitialInterval and mLevelUpInterval
are uninitialized. I have set them to their default values for AI mode.

This patch actually fixes Coverity #76280 while the fix for
Coverity #76280 in fact fixed Coverity #76274. The confusion was due to
the fact that both members were missing in both constructors and
Coverity tracks them as separate issues. Now, both are fixed.