Fix Coverity #76278: Uninitialized member
Member mRunning was uninitialized in the constructor of
KBlocksPlayNetwork. Even though this member, which indicates whether a
game is executed, is always initialized before it is used, it is good
practice to initialize it in the constructor.

It is now set to False in the constructor, indicating that no game is
executed.