Check for NULL pointers
If the replay file can not be loaded, several pointers are set to NULL.
In this case, several member functions crash if called.

At the moment, these functions are not called since main checks if the
replay file has been loaded successfully. However, if the check is
omitted and the replay file cannot be loaded, the program crashes.

I have added checks for NULL pointers in the member functions of
KBlocksRepWin. Now, the program does not crash even if the check is
omitted.