Fix Coverity #76255: Calling risky function
When loading a plain text replay file, fscanf() was used. The use of
this function is discouraged as it can lead to buffer overflows if the
string read is larger than the reserved buffer.

Now, the C-style functions are replaced with their C++ equivalents.
This means the data is read from an ifstream and stored to a string via
an istringstream.