Fix incorrect resize before game is started
When the game window is resized before a game is started, the game scene is not correctly rescaled. When starting the game it might be too small (it seems never to be too large).

This commit fixes the bug by triggering a fitInView() call of the gameView. The same function is triggered by a resize event.

BUGS: 380464
REVIEW: 130167