Do not create KBlocksGraphics and KBlocksSound objects in KBlocksScene
Complex objects being created in constructors is always a big headache
for unit tests, particularly in C++, because constructor calls can't be
avoided and neither can the work be delegated to a virtual member
function.

Instead of creating the objects, KBlocksScene now takes them as input
parameters.

This reduces the number of warnings in the UnitTest-KBlocksWin test
suite to one.