Create SvgItemInterface
To be able to test the next step in the bugfix, I need to verify the
actions taken on a KBlocksSvgItem object. Thus, I want to replace it
with a mock object. To make this possible, I have created a
SvgItemInterface class and replaced all necessary KBlocksSvgItem
pointers with SvgItemInterface pointers.